Extract repository archiver
