+++ src/Konmaripo.Web/Services/IRepositoryArchiver.cs
@@ -1,0 +1,7 @@
+﻿namespace Konmaripo.Web.Services
+{
+    public interface IRepositoryArchiver
+    {
+        RepositoryPath CloneRepositoryWithTagsAndBranches(string repoName);
+    }
+}
